+++ cocos/renderer/terrain/terrain-editor-manage.ts
@@ -1,0 +1,7 @@
+import { TerrainEditor_Mode } from './terrain-editor-mode'
+
+// 管理模式
+//  
+export class TerrainEditor_Manage extends TerrainEditor_Mode
+{
+}
