+++ cocos/renderer/terrain/terrain-editor-mode.ts
@@ -1,0 +1,15 @@
+import { Terrain, TERRAIN_BLOCK_TILE_COMPLEXITY } from './terrain'
+
+export enum eTerrainEditorMode
+{
+    MANAGE,
+    SCULPT,
+    PAINT,
+}
+
+export class TerrainEditor_Mode
+{
+    onUpdate(terrain: Terrain, dtime: number) {
+    }
+}
+
